requirements.txt
PyQT6
requests
python-dotenv
